cloud-trely/requirements.txt
Flask==2.3.2
azure-storage-blob==12.14.1
python-dotenv
